Refactoring to use abstract super class.
