+++ software/contrib/rossler.py
@@ -1,3 +1,4 @@
+from attractor import attractor
 '''
 Implementation of a simple Rossler Attractor, see 
 
@@ -6,55 +7,12 @@ https://en.wikipedia.org/wiki/R%C3%B6ssler_attractor
 Default uses original values of a=0.2, b=0.2, c=5.7
 '''
 
-class rossler:
+class rossler(attractor):
     def __init__(self, point=(5.,5.,5.), params=(0.3,0.2,5.7), dt=0.01):
-        self.x = point[0]
-        self.y = point[1]
-        self.z = point[2]
+        super().__init__(point,dt)
         self.a = params[0]
         self.b = params[1]
         self.c = params[2]
-        self.dt = dt
-        self.x_min = self.x
-        self.y_min = self.y
-        self.z_min = self.z
-        self.x_max = self.x
-        self.y_max = self.y
-        self.z_max = self.z
-        # arbitrary initial range values    
-        self.x_range = 100
-        self.y_range = 100
-        self.z_range = 100
-        
-    # The range of values produced depends on the parameters. If we
-    # know the range, we can then normalise coordinates for use when
-    # generating CV. This method runs through a number of iterations
-    # to estimate ranges.
-    def estimateRanges(self,steps=1000000):
-    
-        # Execute a number of steps to get upper and lower bounds. 
-        for i in range(steps):
-            self.step()
-            
-            self.x_max = max(self.x, self.x_max)
-            self.y_max = max(self.y, self.y_max)
-            self.z_max = max(self.z, self.z_max)
-            self.x_min = min(self.x, self.x_min)
-            self.y_min = min(self.y, self.y_min)
-            self.z_min = min(self.z, self.z_min)
-
-        self.x_range = self.x_max-self.x_min
-        self.y_range = self.y_max-self.y_min
-        self.z_range = self.z_max-self.z_min
-
-    def x_scaled(self):
-        return (100.0 * (self.x - self.x_min))/self.x_range
-
-    def y_scaled(self):
-        return (100.0 * (self.y - self.y_min))/self.y_range
-
-    def z_scaled(self):
-        return (100.0 * (self.z - self.z_min))/self.z_range
     
     def __str__(self):
         return (f"({self.x:2.2f},{self.y:2.2f},{self.z:2.2f})({self.a},{self.b},{self.c})")
